feat: Move "Tips and information" to ticket information screen
