+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_TicketTabNavStack/TicketTabNav_PurchaseTabScreen/TicketTabNav_PurchaseTabScreen.tsx
@@ -11,9 +11,7 @@ import {UpgradeSplash} from './Components/UpgradeSplash';
 import {useRecentFareContracts} from './use-recent-fare-contracts';
 import {FareProductTypeConfig} from '@atb/configuration';
 import {RecentFareContract} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_TicketTabNavStack/TicketTabNav_PurchaseTabScreen/types';
-import {useTipsAndInformationEnabled} from '@atb/stacks-hierarchy/Root_TipsAndInformation/use-tips-and-information-enabled';
 import {useTicketingAssistantEnabled} from '@atb/stacks-hierarchy/Root_TicketAssistantStack/use-ticketing-assistant-enabled';
-import {TipsAndInformationTile} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Assistant/TipsAndInformationTile';
 import {TicketAssistantTile} from '@atb/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Assistant/TicketAssistantTile';
 import {useAnalytics} from '@atb/analytics';
 import {useMobileTokenContextState} from '@atb/mobile-token';
@@ -35,7 +33,6 @@ export const TicketTabNav_PurchaseTabScreen = ({navigation}: Props) => {
   const styles = useStyles();
   const {t} = useTranslation();
 
-  const showTipsAndInformation = useTipsAndInformationEnabled();
   const showTicketAssistant = useTicketingAssistantEnabled();
   const analytics = useAnalytics();
 
@@ -178,14 +175,6 @@ export const TicketTabNav_PurchaseTabScreen = ({navigation}: Props) => {
           />
         )}
 
-        {showTipsAndInformation && (
-          <TipsAndInformationTile
-            onPress={() => {
-              navigation.navigate('Root_TipsAndInformation');
-            }}
-            testID="tipsAndInformation"
-          />
-        )}
         <FareProducts onProductSelect={onProductSelect} />
         {showTicketAssistant && (
           <>
