--- src/stacks-hierarchy/Root_TipsAndInformation/index.ts
@@ -1,1 +1,0 @@
-export {Root_TipsAndInformation} from './Root_TipsAndInformation';
