+++ src/stacks-hierarchy/navigation-types.ts
@@ -113,7 +113,6 @@ type Root_ParkingViolationsConfirmationParams = {
 export type RootStackParamList = {
   NotFound: undefined;
   Root_OnboardingStack: undefined;
-  Root_TipsAndInformation: undefined;
   Root_TicketAssistantStack: undefined;
   Root_TabNavigatorStack: NavigatorScreenParams<TabNavigatorStackParams>;
   Root_LocationSearchByTextScreen: Root_LocationSearchByTextScreenParams;
