+++ src/tips-and-information/index.ts
@@ -1,0 +1,1 @@
+export {TipsAndInformation} from './TipsAndInformation';
