+++ src/translations/screens/subscreens/PurchaseOverview.ts
@@ -256,6 +256,11 @@ const PurchaseOverviewTexts = {
         'Ticket description',
         'Billettbeskriving',
       ),
+      tipsInformation: _(
+        'Generelt om billetter',
+        'About tickets in general',
+        'Generelt om bilettar',
+      ),
     },
   },
 };
